move delete/rdelete cmd tests to own modules
